Workaround for Issue #5
This defaults the deck ease to 2500 in case, such as with filtered decks, the deck has no configuration setttings.

Ideal would be to look back to that card's home deck, but this workaround should work for now -- the ease factors should correct themselves as soon as you're back to regular scheduling.